shuffle routines between Reference and SystemInformation
